Moving dummy auth to Common package
