+++ Demo/SwizlyPeasy.Demo.API/Extensions/AuthorizationExtensions.cs
@@ -1,7 +1,4 @@
-﻿using System.Net;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using SwizlyPeasy.Demo.API.Authorization;
 
 namespace SwizlyPeasy.Demo.API.Extensions;
@@ -9,29 +6,16 @@ namespace SwizlyPeasy.Demo.API.Extensions;
 public static class AuthorizationExtensions
 {
     /// <summary>
-    ///     Adding demo authorization policy, checking if user's uid/sub matches bob's one
+    ///     Test policy, checking if user is bob
     /// </summary>
     /// <param name="services"></param>
-    public static void SetAuthenticationAndAuthorization(this IServiceCollection services)
+    public static void SetAuthorization(this IServiceCollection services)
     {
-        services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme)
-            .AddCookie(CookieAuthenticationDefaults.AuthenticationScheme, options =>
-            {
-                options.Events.OnRedirectToAccessDenied = UnAuthorizedResponse;
-                options.Events.OnRedirectToLogin = UnAuthorizedResponse;
-            });
-
         services.AddAuthorization(options =>
         {
             options.AddPolicy("AreYouBob", policy =>
                 policy.Requirements.Add(new BobRequirement()));
         });
         services.AddSingleton<IAuthorizationHandler, AppAuthorizationHandler>();
-    }
-
-    private static Task UnAuthorizedResponse(RedirectContext<CookieAuthenticationOptions> context)
-    {
-        context.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
-        return Task.CompletedTask;
     }
 }
